Create Livechat_v0.2
This version did the job.
(1) Removes all stopwords from livechate export
(2) Counts the frequency of the remaining words
(3) Exports the word plus frequency into a CSV file which I can be used in Power BI

Drawbacks:
(1) The files names for input and output are hardcoded
(2) The output is inserted into an existing file that has to be specified rather than creating a new CSV from scratch and naming it.
(3) You can specify the individual sheet in the Livechat export file. It assumes its called sheet 1. 
(4) There is no basic GUI - not sure how you run this from the command prompt. Need something easier for other users.